pyproject.toml
[project]
name = "cartoon-pipeline"
version = "0.1.0"
description = "Multi-agent AI cartoon production pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core LLM / agent framework
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "ollama>=0.3.0",
    # Web UI backend
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "python-multipart>=0.0.9",
    # Data / config
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    # Trend research
    "requests>=2.31.0",
    "feedparser>=6.0.0",
    # YouTube upload (optional)
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.1.0",
    # TTS (optional - install separately if needed)
    # "TTS>=0.22.0",  # Coqui - heavy, uncomment if wanted
]

[project.optional-dependencies]
tts = ["TTS>=0.22.0"]
dev = ["ruff>=0.6.0", "pytest>=8.0.0", "httpx>=0.27.0"]

[project.scripts]
cartoon = "main:main"
cartoon-ui = "webui.server:start"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
